cleanup based on HP-UX CC -k signal option

+++ src/ftpget.c
@@ -1,5 +1,5 @@
 /*
- * $Id: ftpget.c,v 1.34 1996/08/30 23:23:30 wessels Exp $
+ * $Id: ftpget.c,v 1.35 1996/09/12 03:24:04 wessels Exp $
  *
  * DEBUG: section 0     FTP Retrieval
  * AUTHOR: Harvest Derived
@@ -1492,7 +1492,7 @@ state_t do_port(r)
 	(naddr >> 16) & 0xFF,
 	(naddr >> 8) & 0xFF,
 	naddr & 0xFF,
-	(port >> 8) & 0xFF,
+	((int) port >> 8) & 0xFF,
 	port & 0xFF);
     SEND_CBUF;
